refactor(advisor)!: Move advisor configuration classes to advisor module
Since the previous commit the model does not require the advisor
specific configuration classes anymore. Move them to the advisor module
to prepare for moving the advice provider implementations to their own
plugin modules.
